Feat: adedd the logic to register and login users and all security configuration

+++ Backend/MunchEZ/src/main/java/mx/com/MunchEZ/MunchEZ/domain/personal/DataPersonalUpdate.java
@@ -2,5 +2,5 @@ package mx.com.MunchEZ.MunchEZ.domain.personal;
 
 import mx.com.MunchEZ.MunchEZ.domain.Role.Role;
 
-public record DataPersonalUpdate(String name, Boolean active, int roleId, String phone) {
+public record DataPersonalUpdate(String name, Boolean active, Long roleId, String phone) {
 }
